added a filter for author name.
To add more filters, use the search-filter tag, currently supporting only text and boolean fields.
Make sure that the group and update attributes are the same (they are intrinsic to the level search, not the filter tag itself, so keep them the same)
The name attribute is the database field that shall be searched (e.g. name, author, rating, etc)